Restore persistent settings after Chromium update
Iiiiinteresting.

Looks like I misinterpreted the `sendResponse` parameter in
`chrome.runtime.sendMessage`. I assumed it was a callback to be
triggered manually. Turns out it's triggered automatically. This was
leading to all sorts of weird behavior when opening and closing the
options page. Especially bad: the "Theme defaults to" option stopped
working.

I didn't catch this sooner because it wasn't a problem until Chromium
[fixed their sendResponse implementation](https://chromium.googlesource.com/chromium/src/+/3607f1773518f29bc704daf7de5a89961c2c4730%5E!/).
Heh.

It's a simple fix. We just send the callback in the `request` object
instead. Not going to lie, though: debugging this had me scratching
my head for a little while :P